// resources/js/types/models/others.ts
export interface Prerequisite {
    entity_type: string;
    entity_id: number;

    prerequisite_type: PrerequisiteType;
    prerequisite_id?: number | null;

    // nullableMorphs('entitytable')
    entitytable_type?: string | null;
    entitytable_id?: number | null;

    description?: string | null;

    is_mandatory: boolean;

    min_score?: number | null;
    min_attempts?: number | null;

    logic: LogicOperator;
}

export interface CustomMapDataProposal {
    id: number;
    title: string;
}